Move GarudStats from deprecated header into new header.
