+++ Sequence/summstats/Makefile.am
@@ -1,4 +1,4 @@
 pkgincludedir=$(prefix)/include/Sequence/summstats
 
 pkginclude_HEADERS = classics.hpp thetapi.hpp thetaw.hpp thetah.hpp thetal.hpp auxillary.hpp nvariablesites.hpp allele_counts.hpp \
-					 util.hpp ld.hpp nsl.hpp
+					 util.hpp ld.hpp nsl.hpp garud.hpp
